moved renderDiffView into it's own component, prepping for diff fetching
